Move cfg classes to separat module and implement shallow copy for global state
