+++ mythril/laser/ethereum/svm.py
@@ -1,14 +1,13 @@
 from mythril.laser.ethereum import helper
 from ethereum import utils
 from enum import Enum
-from flags import Flags
 from z3 import *
 import binascii
 import copy
 import logging
 import re
 from mythril.laser.ethereum.state import *
-
+from mythril.laser.ethereum.cfg import *
 TT256 = 2 ** 256
 TT256M1 = 2 ** 256 - 1
 
@@ -20,80 +19,9 @@ class CalldataType(Enum):
     SYMBOLIC = 2
 
 
-class JumpType(Enum):
-    CONDITIONAL = 1
-    UNCONDITIONAL = 2
-    CALL = 3
-    RETURN = 4
-
-
 class SVMError(Exception):
     pass
 
-
-
-'''
-The final analysis result is represented as a graph. Each node of the graph represents a basic block of code.
-The states[] list contains the individual global state at each program counter position. There is one set of constraints on each node.
-A list of edges between nodes with associated constraints is also saved. This is not strictly necessary for analysis, but is useful
-for drawing a nice control flow graph.
-'''
-
-
-class NodeFlags(Flags):
-    FUNC_ENTRY = 1
-    CALL_RETURN = 2
-
-
-class Node:
-
-    def __init__(self, contract_name, start_addr=0, constraints=[]):
-        self.contract_name = contract_name
-        self.start_addr = start_addr
-        self.states = []
-        self.constraints = constraints
-        self.function_name = "unknown"
-        self.flags = NodeFlags()
-
-        # Self-assign a unique ID
-
-        global gbl_next_uid
-
-        self.uid = gbl_next_uid
-        gbl_next_uid += 1
-
-    def get_cfg_dict(self):
-
-        code = ""
-
-        for state in self.states:
-
-            instruction = state.get_current_instruction()
-
-            code += str(instruction['address']) + " " + instruction['opcode']
-            if instruction['opcode'].startswith("PUSH"):
-                code += " " + instruction['argument']
-
-            code += "\\n"
-
-        return {'contract_name': self.contract_name, 'start_addr': self.start_addr, 'function_name': self.function_name, 'code': code}
-
-
-class Edge:
-
-    def __init__(self, node_from, node_to, edge_type=JumpType.UNCONDITIONAL, condition=None):
-
-        self.node_from = node_from
-        self.node_to = node_to
-        self.type = edge_type
-        self.condition = condition
-
-    def __str__(self):
-        return str(self.as_dict())
-
-    def as_dict(self):
-
-        return {'from': self.node_from, 'to': self.node_to}
 
 '''
 Main symbolic execution engine.
